// src/app/modules/staff/staff-admissions/model/job-application-comments-view.ts
/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * JobApplicationCommentsView
 */
export interface JobApplicationCommentsView { 
    /**
     * Gets or sets JobApplicationId
     */
    jobApplicationId?: string;
    /**
     * Gets or sets ProcessedComments
     */
    processedComments?: string;
}