// src/app/modules/staff/staff-admissions/model/header-values-view.ts
/**
 * Teacher API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface HeaderValuesView {
    fullName?: string;
    mobileNumber?: string;
    primaryEmail?: string;
    applicationReferenceNo?: string;
    resumeFile?: string;
    statusCode?: string;
    processedComments?: string;
    fileType?: string;
    seekingPosition?: string;
}